requirements.txt
cycler==0.11.0
fonttools==4.37.1
gitdb==4.0.9
GitPython==3.1.29
kiwisolver==1.4.4
matplotlib==3.5.3
networkx==2.8.6
numpy==1.23.3
packaging==21.3
Pillow==9.2.0
pkg_resources==0.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
scipy==1.9.1
six==1.16.0
smmap==5.0.0
